bookfarm_api/requirements.txt
asgiref 
atomicwrites==1.4.0
attrs==21.2.0
boto3 
botocore 
brotlipy==0.7.0
certifi==2021.5.30
cffi 
colorama==0.4.4
configparser 
cryptography 
Django 
django-cors-headers 
djangorestframework 
idna 
iniconfig==1.1.1
jmespath 
packaging==20.9
pluggy==0.13.1
psycopg2 
py==1.10.0
pycparser 
PyJWT 
pyOpenSSL 
pyparsing==2.4.7
PySocks 
pytest==6.2.4
python-dateutil 
pytz 
s3transfer 
six 
sqlparse 
toml==0.10.2
urllib3 
win-inet-pton 
wincertstore==0.2
